various: move hwclock and rc.local files to /etc
It makes sense that /etc/rc.local and /etc/rc.shutdown is put in the
main /etc directory. Besides, if dinit-rc isn't installed, it won't
conflict with anything else (including openrc, s6-scripts, runit-rc,
etc). It's a matter of packaging, to be honest.

For hwclock, however, in anticipation of esettingsd, the non-OpenRC
inits need to agree on where to put hwclock.conf file, so I
pre-emptively put it on /etc so esettingsd can read it.